refactor(graphics): restructure graphics feature into layered architecture (#14)
