+++ src/app/features/graphics/components/hours-by-weekday-vehicle-chart/hours-by-weekday-vehicle-chart.spec.ts
@@ -1,10 +1,12 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { HttpClientModule } from '@angular/common/http';
+import { Auth } from '@angular/fire/auth';
 
 import { HoursByWeekdayVehicleChartComponent } from './hours-by-weekday-vehicle-chart';
-import { HttpClientModule } from '@angular/common/http';
-import { GraphicsServices } from '../../services/graphics-services';
+
+import { GraphicsServices } from '../../data-access/graphics-services';
 import { VehicleService } from '../../../vehicle/services/vehicle-service/vehicle-service';
-import { Auth } from '@angular/fire/auth';
+
 
 export const authMock = {
   currentUser : {
