+++ src/app/features/graphics/components/most-used-vehicle-chart/most-used-vehicle-chart.ts
@@ -3,7 +3,7 @@ import Chart from 'chart.js/auto';
 
 import { TimePeriod } from '../../enums/time-period.enum';
 
-import { GraphicsServices } from '../../services/graphics-services';
+import { GraphicsServices } from '../../data-access/graphics-services';
 import { VehicleService } from '../../../vehicle/services/vehicle-service/vehicle-service';
 @Component({
   selector: 'app-most-used-vehicle-chart',
@@ -19,8 +19,8 @@ export class MostUsedVehicleChartComponent implements OnDestroy {
 
   public period = input<TimePeriod>(TimePeriod.Month);
   
-  private graphicsService = inject(GraphicsServices);
-  private vehicleService = inject(VehicleService);
+  private readonly graphicsService = inject(GraphicsServices);
+  private readonly vehicleService = inject(VehicleService);
 
   private chart!: Chart;
