+++ src/app/features/graphics/components/vehicle-usage-hours-chart/vehicle-usage-hours-chart.spec.ts
@@ -4,7 +4,7 @@ import { Auth } from '@angular/fire/auth';
 
 import { VehicleUsageHoursChartComponent } from './vehicle-usage-hours-chart';
 
-import { GraphicsServices } from '../../services/graphics-services';
+import { GraphicsServices } from '../../data-access/graphics-services';
 import { VehicleService } from '../../../vehicle/services/vehicle-service/vehicle-service';
 
 export const authMock = {
