+++ src/app/features/graphics/components/vehicle-usage-hours-chart/vehicle-usage-hours-chart.ts
@@ -1,7 +1,8 @@
 import { Component, effect, ElementRef, inject, input, OnDestroy, ViewChild } from '@angular/core';
 import Chart from 'chart.js/auto';
 import { TimePeriod } from '../../enums/time-period.enum';
-import { GraphicsServices } from '../../services/graphics-services';
+
+import { GraphicsServices } from '../../data-access/graphics-services';
 import { VehicleService } from '../../../vehicle/services/vehicle-service/vehicle-service';
 
 @Component({
@@ -18,8 +19,8 @@ export class VehicleUsageHoursChartComponent implements OnDestroy {
 
   public period = input<TimePeriod>(TimePeriod.Month);
   
-  private graphicsService = inject(GraphicsServices);
-  private vehicleService = inject(VehicleService);
+  private readonly graphicsService = inject(GraphicsServices);
+  private readonly vehicleService = inject(VehicleService);
 
   private chart!: Chart;
