+++ src/app/features/graphics/pages/graphics/graphics.spec.ts
